Return early in LinksDb.add if slot is occupied
Otherwise we will increment the link count even when we did not add a
link, which will screw up the link database. Before, I did `printAndExit`
here but I removed that to fix another bug, but not doing anything
ain't it either: we just need to exit the function, not the program!

To avoid problems like this in the future, check that the link database
link count and the actual number of links match in `verify`. Also, run
`verify` passes in ReleaseSafe, since that will cause crashes if the
database state is messed up, which will make it obvious that something
is wrong and hopefully inspire the user to make a bug report instead of
proceeding with a messed up database causing inscrutable issues down the
line. Also also, fix `verify` by actually adding links to the hash set
when checking for dupes...

Thanks to cryptocode on the Zig Discord for reporting this issue and
providing an excellent repro.